% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardianr.R
\name{get_guardian}
\alias{get_guardian}
\title{Search Guardian API for news articles that match the criteria}
\usage{
get_guardian(keywords, section = NULL, from, to, api_key,
  verbose = TRUE)
}
\arguments{
\item{keywords}{Keyword to search Guardian API. Example: "Thatcher". For
multiple keywordsuse "Margaret+Hilda+Thatcher".}

\item{section}{Specifies news sections to narrow the query or NULL to search
everywhere.}

\item{from, to}{Start and end date of search (both are included in the
search).}

\item{api_key}{A Guardian API-key is necessary to retrieve the full text of
news articles. A Guardian API-key can be obtained by registering at
<http://open-platform.theguardian.com/access/>}

\item{verbose}{Should progress and other messages be displayed (logical).}
}
\value{
data.frame with all variables delivered by the API
}
\description{
The function get_guardian takes four variables (keyword(s),
  starting date, end date, and API-key) and returns a data frame with a
  column for every variable returned by the API columns, with the last column
  including the full text of the article.

  Search criteria accepts single or multiple keywords. It also accepts
  Boolean queries with AND/OR/NOT between words to refine searches. For exact
  phrasesand matches, please encapsulate the keywords in double quotes or %22
  (e.g "%22Death+of+Margaret+Thatcher%22").

  From version 0.5 onwards, the function get_guardian returns the full text
  of articles and requires a Guardian API-key. Guardian API-key can be
  obtained by registering at <http://open-platform.theguardian.com/access/>.
}
\examples{
\dontrun{
results <- get_guardian(
    keywords = "Theresa May AND (Brexit OR EU)",
    from = "2019-01-16",
    to = "2019-01-30",
    api_key = "212d23d3-c7b2-4273-8f1b-289a0803ca4b"
)
}
}
