\name{parse_json_to_df.nokey}
\alias{parse_json_to_df.nokey}
\title{
Parses json data to data frame
}
\description{
The function parse_json_to_df parses the json data to a data frame in R. The function is called by get_guardian and should not be used standalone.
}
\usage{
parse_json_to_df.nokey(api.responses.nokey)
}
\arguments{
  \item{api.responses.nokey}{
Parses JSON to data frame 
}
}
\author{
Marco Toledo Bastos & Cornelius Puschmann
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ IO }
\keyword{ utilities }
