\name{parse_json_to_df}
\alias{parse_json_to_df}
\title{
Parses json data to data frame
}
\description{
The function parse_json_to_df parses the json data to a data frame in R. The function is called by get_guardian and should not be used standalone.
}
\usage{
parse_json_to_df(api.responses)
}
\arguments{
  \item{api.responses}{
JSON response from Guardian API
}
}
\author{
Marco Toledo Bastos & Cornelius Puschmann
}
\keyword{ IO }
\keyword{ utilities }
